requirements.txt
# Core Framework
streamlit>=1.28.0

# LangChain packages
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.1
langchain-groq>=0.1.0
langchain-text-splitters>=1.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0

# Vector store
chromadb>=0.4.22

# Embeddings (local, no API key needed)
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2
